backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9

# Document processing
pdfplumber>=0.11.0
pdf2image>=1.17.0
paddleocr>=2.7.0
paddlepaddle>=2.6.0
python-docx>=1.1.0

# NLP
spacy>=3.7.0
transformers>=4.40.0
sentence-transformers>=2.7.0
torch>=2.2.0

# Graph
neo4j>=5.19.0

# Vector
faiss-cpu>=1.8.0

# LLM
groq>=0.9.0

# Cache
redis>=5.0.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
hypothesis>=6.100.0
